Center crop

+++ preprocessing_data.py
@@ -5,15 +5,31 @@ import os, shutil, pathlib
 # External
 import cv2
 import pandas as pd
+import splitfolders
 
-HEIGHT = 224
-WIDTH = 224
+CROP_HEIGHT = 600
+CROP_WIDTH = 600
+CENTER_CROP = False
+SCALE_HEIGHT = 224
+SCALE_WIDTH = 224
 IMAGE_FOLDER = "scaled_images"
 CONTINENTS = ["Europe", "Asia"]
 LIMIT_AMOUNT = True
 IMAGE_COUNT = {"Europe":8000,"Asia":8000}
 SEED = 42
+ONLY_SPLIT =True
 
+# center crops images to maximum possible height and width within the specified values
+def center_crop(img, dim):
+	width, height = img.shape[1], img.shape[0]
+
+	# process crop width and height for max available dimension
+	crop_width = dim[0] if dim[0]<img.shape[1] else img.shape[1]
+	crop_height = dim[1] if dim[1]<img.shape[0] else img.shape[0] 
+	mid_x, mid_y = int(width/2), int(height/2)
+	cw2, ch2 = int(crop_width/2), int(crop_height/2) 
+	crop_img = img[mid_y-ch2:mid_y+ch2, mid_x-cw2:mid_x+cw2]
+	return crop_img
 
 # resizes images to Height and Width dimensions
 def resize_sample(row: pd.Series) -> str:
@@ -22,7 +38,9 @@ def resize_sample(row: pd.Series) -> str:
     image_name = row["image_name"]
 
     image = cv2.imread(image_path)
-    scaled_image = cv2.resize(image, (WIDTH, HEIGHT))
+    if CENTER_CROP:
+        image = center_crop(image, (600,600))
+    scaled_image = cv2.resize(image, (SCALE_WIDTH, SCALE_HEIGHT))
 
     path = f'{IMAGE_FOLDER}/{continent}'
     if not os.path.exists(path):
@@ -49,18 +67,21 @@ def get_filtered_dataframe(df: pd.DataFrame):
     return df_result
 
 def main():
+    if not ONLY_SPLIT:
+        # delete scaled_images folder
+        if os.path.exists(IMAGE_FOLDER):
+            shutil.rmtree(IMAGE_FOLDER)
+        os.mkdir(IMAGE_FOLDER)
 
-    # delete scaled_images folder
-    if os.path.exists(IMAGE_FOLDER):
-        shutil.rmtree(IMAGE_FOLDER)
-    os.mkdir(IMAGE_FOLDER)
-
-    df_metadata = pd.read_csv('country_data.csv')  # import metadata from append_metadata.py, includes all countries and continents
-    df_filtered_continent = get_filtered_dataframe(df_metadata)
-    #df_metadata = df_metadata.iloc[:10, :] # for testing
-    df_resized = df_filtered_continent
-    df_resized["path"] = df_filtered_continent.apply(lambda sample: resize_sample(sample), axis=1)
-    df_resized.to_csv("preprocessing_resized.csv", index=False)
+        df_metadata = pd.read_csv('country_data.csv')  # import metadata from append_metadata.py, includes all countries and continents
+        df_filtered_continent = get_filtered_dataframe(df_metadata)
+        #df_metadata = df_metadata.iloc[:10, :] # for testing
+        df_resized = df_filtered_continent
+        df_resized["path"] = df_filtered_continent.apply(lambda sample: resize_sample(sample), axis=1)
+        df_resized.to_csv("preprocessing_resized.csv", index=False)
+        
+    else:
+        splitfolders.ratio("scaled_images",seed=42, output="scaled_images_splitted", ratio=(0.7, 0.2, 0.1))
 
 if __name__ == "__main__":
     main()
